#feature Batching of sb.User selects/inserts
Now provides `sb.User.getMultiple`, which smartly gets a list of users,
either from cache, or from database, or combined. Does not support addition
of new users.

This is to avoid fetching users 1-by-1 when it is known multiple will be
used at a single time.

Insertion of sb.User-s is now also batched. This is to improve performance
by quite a bit when a lot of new users are being encountered.